Campaign Server, DB
Not quite there, but some of the basic stuff setup for the server
and the DB layer. For now we'll stick just with raw Ruby objects
and persisting to disk directly in a marshalled form. This way each
"server" can manage it's own shard. This should keep things a
little "simpler" when we do all the rebalancing stuff and what not.

+ Added db.rb (database layer)
+ Set everything in the Gemfile (currently) to be in the default group
+ Added two campaign processes to the Procfile (foreman)
+ Created the campaign server
+ Created the campaign server rackup file (campaign-config.ru)